Create Draw Panel
This project consists of a total of 5 classes. The goal of the project is to write a program that will draw three types of shapes. The shapes drawn by the program will be lines, ovals, and rectangles. The program randomly generates 3 values that represent the number of lines, ovals, and rectangles to be drawn. The program also generates a random color for each shape drawn. Class MyLine builds a line using 4 coordinates. Likewise, classes MyOval and MyRectangle build ovals and rectangles. 
Each type of shape generated by the program is stored in a separate array. The specific values that build each shape are specified in the code above. The program generates random values for the 4 coordinates. Thus, the size and position of each shape generated is random. The class above calls methods declared and defined in classes MyLine, MyOval, and MyRectangle.