Create Assignment 2: MyLine
In addition to the description given in file Draw Panel, the code above shows the syntax for performing validation in the set methods. For setting coordinate values, if the value of the argument supplied is less than 0, value of the argument is set to 0. Notice that the class's instance variable have private access, so set methods are needed to manipulate the values of the instance variables, and get methods are needed to retrieve their values. The class presented here declares and defines a method called drawLine to set the color of the shape, and to draw the shape.