Create MyOval
File Draw Panel briefly describes the purpose of this class. File MyLine from this assignment also gives an explanation of some of the syntax that is presented in this class. Classes MyRectangle and MyOval have an additional instance variable that can be set to true or false. Initially, the value of the instance variable is false. The value becomes true after the statements in method drawOval are executed indicating that the shape has been filled with color. The drawOval, and fillOval method of class Graphics takes 4 arguments. These arguments are the x and y values of the upper left corner of the shape. Also the height and width of the shape. The class above declares 4 separate methods to find these values.