Create MyRectangle Assignment 2
File Draw Panel briefly describes the purpose of this class. File MyLine from this assignment also gives an explanation of some of the syntax that is presented in this class. Classes MyRectangle and MyOval have an additional instance variable that can be set to true or false.
The drawRect and fillRect methods of class Graphics also take the same parameters as methods drawOval and fillOval.