Create Assignment 3: MyShape
Since MyShape is the parent class of classes MyLine, MyRectangle, and MyOval, class MyShape will contain data that is common to all the child classes. Class MyShape is declared abstract because it contains abstract methods. Set and get methods are also declared and defined in this class. For example, the class contains an instance variable of type Color called color. The get method's return type is of type Color.