Create Draw Panel 1
Class DrawPanel1 creates an array of type MyShape. The user of this application will input the number of shapes the program should draw. This user input is the number of indexes in the array. The program also receives a second input value from the user. The user decides if the shapes drawn should be filled with color, and which shapes should be filled with color. In order to do this, an array named choices is created. Class JOptionPane will present the user with a list of options to select from. 
This program generates random colors to draw/fill the shapes. Also, the program generates random numbers for the coordinate values. The output of the program will display shapes with random sizes and in random positions. The random numbers are generated using class SecureRandom. 
This code presents the concept of polymorphism where creating an instance of the child class is set equal to a reference variable with type of parent class. 
This class also contains a method that will be used to display a label on JPanel. The method outputs the number of lines, rectangles, and ovals drawn by the program. Since ovals and rectangles can be filled with color, classes MyOval and MyRectangle contain instance variables that can be set to true or false. The set methods that set these instance variables are declared static. Hence these set methods can be accessed from outside the class. Class DrawPanel1 calls these set methods.