Create MyLine
This class contains calls to the superclass constructors to initialize variables contained in the superclass. Since this is a child class of abstract class MyShape, this class must implement all abstract methods from the parent class. If the abstract methods are not implemented here, this class also needs to be declared abstract. The purpose of this class is to build a line.