Create Assignment 3: My Oval 1
This class contains calls to the superclass constructors to initialize variables contained in the superclass. Since this is a child class of abstract class MyShape, this class must implement all abstract methods from the parent class. If the abstract methods are not implemented here, this class also needs to be declared abstract. The purpose of this class is to build a line.   
This class has declared an instance variable called ovalFilledFlag that is not applicable to all 3 child classes. This is boolean variable that indicates when a shape has been filled. This class declares and defines 4 methods whose values are needed for methods drawOval and fillOval. The purpose of this class is to build an oval.