Create Assignment 3: Test Draw Shapes1
The main method in this class is where the program will begin executing. The code above creates an instance of class DrawPanel1, and JFrame. In addition, the class creates an instance of class JLabel. As an argument to the constructor of JLabel, method shapeStatusText is called. The method shapeStatusText is declared and defined in class DrawPanel1.