Create Bulls Eye Game Board Pattern
The program draws a total of 5 circles. Each consecutive smaller circle is drawn inside the larger circle. The diameter of each consecutive smaller circle decreases by a uniform amount. Consecutive circles alternate in color. The program randomly generates the two colors that consecutive circles will alternate between. The circles are drawn in the center of the JPanel container. Methods of class Graphics are utilized to draw and fill the circles.