Create Assignment 5: Drawing concentric circles
The program above draws 12 concentric circles. The circles are drawn in the center of JPanel. Class Circles is a direct subclass of class JPanel. The diameter of consecutive circles is decreased by a uniform amount. Variable called shift is used to control where the upper-left x, and y coordinates of the bounding rectangle will be placed in reference to the center of JPanel. For each smaller circle the value of shift changes because each circle has different upper left coordinates of the rectangle bounding the circle.