Create Input from user
Program above asks the user to enter an integer greater than 0. If the integer entered is less than or equal to 0 then program gives the prompt again and allow the user to enter another value . The program will keep giving out this prompt until user inputs an integer greater than 0. If and when the user enters an integer greater than 0, the program will output a message and then end.