Create class Card
The goal of this program is to identify if there is a 'full house', 'straight', 'flush', pairs of cards, and a specific type of card in a hand of 5 cards. Additional classes are developed to achieve this goal. Made my own modifications to this program presented by Paul and Harvey Deitel. 
The constructor initializes variables face and suit. The toString method assigns String values to specific cards. This method returns the face and suit of the card. The getFace and getSuit methods return the face and suit respectively.