Create Deck of Cards
The constructor initializes variables faces and suits. This program presents an array of data type Card and the name of the array is deck. Each element in array 'deck' consists of a value from array faces and a value from the array called suits. Method shuffle shuffles the cards. Method dealCard deals out a card. I've written 7 methods which can give an analysis of what is included in the hand of 5 dealt cards.