Create Assignment 11: Guess-the-number game
This file contains my code as well as system generated code. System generated code gets created when the Graphical User Interface is designed. For example, code gets generated when containers such as JFrame, and JPanel, and JTextFields are added to the GUI. Another example of when code gets generated is when JLabels are added to the form, and when the the font of a JLabel is set. 
This application is designed to do the following. The program generates a random number in the range of 1-1000 that's to be guessed by the game user. The user has an infinite number of tries to guess the number. If the user guesses the correct answer, the game user is informed that their answer is correct. If a wrong answer is guessed, the user is informed that their guess was incorrect and whether their guess was too low or too high. The GUI background color changes to red if the user's new guess is closer to the correct answer than the previous guess. The background color changes to blue if the user's new guess is further away from the correct answer than the previous guess. When the user finally guesses the correct answer, the number of tries taken is displayed. The user can press the red X in the upper right corner of the window at anytime to terminate the game.  
I programmed a button on the GUI that the user can press to start and play the game over again. In this case, program generates a new random number. 
Incorporated a JTextField to allow for user input that becomes uneditable if the user guesses the correct answer. Also, incorporated JLabels to give messages to the user depending on user response. 
In the code I included set and get methods for class instance variables. 
I made use of class ArrayList and class Color. Also made use of try/catch syntax for handling errors. 
When user guesses the correct answer, program displays a label with an icon. In order to do this, an ImageIcon was added to a JLabel. I used an object of class ImageIcon. Then, I added the JLabel to the JFrame. Program created objects of classes JLabel, and JTextField. Program validates user entry. If user does not enter a numeric value, message is displayed  stating the entry must be in numeric form. If user entry is outside the range of 1-1000, message is displayed stating that guess should be within that range.      
File includes comments to explain the code.