Create Assignment 12: Resolving Deadlock (Class Producer1)
This program will execute the Producer1 thread and this Producer1 thread will add  a value to the ArrayBlockingQueue called buffer1. The run method in this class calls method blockingPut of interface Buffer. The blockingPut method adds a value to ArrayBlockingQueue called buffer1. The implemented blockingPut method shows that the Producer1 thread waits if buffer11already holds a value and its occupied.