Create Assignment 12: Resolving Deadlock (Class Consumer)
The thread of class Consumer reads and removes values from the ArrayBlockingQueues declared in class BlockingBuffer. If there aren't any values in the ArrayBlockingQueues then the Consumer thread waits until values become available in the ArrayBlockingQueues. Specifically, the blockingGet method does not come to completion until values become available in the ArrayBlockingQueues. The blockingGet method called from the run method retrieves and removes the values from both ArrayBlockingQueues. The blockingGet method will not come to completion until both values are removed from the ArrayBlockingQueues. This program also outputs the sum of the values read and retrieved.