Create Assignment 13: Generic Classes and Methods (class Stack<T>)
In this class a new object of class ArrayList is created. I added 2 more methods to this class: method peek, and method display. Method peek outputs the very top value of a Stack. Method display outputs all the values of the Stack. The peek method may invoke constructor of class EmptyStackException. 
Since class Stack is declared as a generic class, an object of class Stack can be created that can hold values of type Integer. Another object of class Stack can be created that can hold values of type Double.