Create Assignment 13: Generic Classes and Methods(class StackTest2)
Two arrays are declared in this class. The first array holds values of type Double and the other array holds values of type Integer. This class calls the testPush method twice. The first call to testPush takes different arguments than the second call to testPush. The 1st argument of testPush needs to be a value of type String. The second argument of method testPush needs to be an object of class Stack. Class Stack has a type associated with it like class ArrayList. The 3rd argument needs to be an array of a specific type. Method testPush calls method push of class Stack. Method push adds a value to the ArrayList. So method testPush takes all the values from the array and adds them to the ArrayList. Method testPush is called once to take values from an array of type Integer and add to those values to an ArrayList of type Integer. Method testPush is called the second time to take values from an array of type Double and add those values to an ArrayList of type Double. Method testPush prints out the values that are going to be added to the ArrayList. 
After adding values to the ArrayList, calls to methods peek and display are included. This  class outputs the return value of method peek. This class also outputs the return value of method display. 
Next we include a call to method testPop. Method testPop includes a call to method pop. Method pop removes values from the ArrayList. Method testPop takes 2 arguments. The 1st argument is a value of type String. The second argument is an object of class Stack. Method testPop prints out the values that are removed from the ArrayList. This class calls method testPop twice. Method testPop is called the first time to remove values from an ArrayList of type Double. Method testPop is called the second time to remove values from an ArrayList of type Integer. A call to method testPop is included after calls to methods peek and display. When the testPop method is done removing all of the values from the ArrayList, it throws an exception. This thrown exception is an indication that all the values have been removed from the ArrayList. This file contains comments that explains the code.